client-hints/feature-policy-disable-cross-origin-ch.https.sub.html

<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>

promise_test(t => {
  
  return fetch("/client-hints/echo_client_hints_received.py").then(r => {
      assert_equals(r.status, 200);
      assert_false(r.headers.has("dpr-received"));
    })
  },
  "Disabled same-origin Client Hints with Feature Policy; hint was not sent." 
);

</script>

</body>
</html>
